whisperdesk/requirements.txt
PyQt6>=6.5.0
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
torch>=2.0.0
